+++ lib/trigger.dart
@@ -27,18 +27,18 @@ abstract class Trigger {
   }
 
   final _valueTable = <String, Object>{};
-  final _stateTable = <String, Set<_TriggerState>>{};
+  final _stateTable = <String, Set<TriggerState>>{};
 
-  void _register(List<String> keys, _TriggerState state) {
+  void _register(List<String> keys, TriggerState state) {
     for (var key in keys) {
       if (!_stateTable.containsKey(key)) {
-        _stateTable[key] = <_TriggerState>{};
+        _stateTable[key] = <TriggerState>{};
       }
       _stateTable[key]?.add(state);
     }
   }
 
-  void _unRegister(_TriggerState state) {
+  void _unRegister(TriggerState state) {
     for (var states in _stateTable.values) {
       states.remove(state);
     }
@@ -46,28 +46,31 @@ abstract class Trigger {
 
   ///Check if the key is existed in _stateTable, makes an empty list of _TriggerState if it's not.
   void _ensureStateKey(String key) {
-    if (!_stateTable.containsKey(key)) _stateTable[key] = <_TriggerState>{};
+    if (!_stateTable.containsKey(key)) _stateTable[key] = <TriggerState>{};
   }
 
+  ///Get the according value.
   dynamic getValue(String key) => _valueTable[key];
 
+  ///To set the according value and update the states which listen to this value.
   void setValue(String key, Object val) {
     _valueTable[key] = val;
     _ensureStateKey(key);
     for (var state in _stateTable[key]!) {
-      state._update();
+      state.update();
     }
   }
 
+  ///To set the according values and update the states which listen to thiese values.
   void setMultiValue(Map<String, dynamic> maps) {
-    Set<_TriggerState> _statesToUpdate = {};
+    Set<TriggerState> _statesToUpdate = {};
     for (var key in maps.keys) {
       _ensureStateKey(key);
       _valueTable[key] = maps[key];
       _statesToUpdate.addAll([...?_stateTable[key]]);
     }
     for (var state in _statesToUpdate) {
-      state._update();
+      state.update();
     }
   }
 }
@@ -82,4 +85,18 @@ class TriggerError implements Exception {
 
   @override
   String toString() => cause ?? '';
+}
+
+abstract class TriggerState<T extends Trigger, STF extends StatefulWidget> extends State<STF> {
+  List<String> get listenTo;
+  T get trigger => Trigger.of<T>();
+  void update() {
+    setState(() {});
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    trigger._register(listenTo, this);
+  }
 }